+++ packages/redux_persist_web/lib/redux_persist_web.dart
@@ -3,16 +3,14 @@ import 'dart:html';
 
 import 'package:redux_persist/redux_persist.dart';
 
-/// Storage engine to use with Flutter
+/// Storage engine to use with Web
 class WebStorage implements StorageEngine {
   final String key;
 
   WebStorage(this.key);
 
   @override
-  load() {
-    return new Future.value(window.localStorage[key]);
-  }
+  load() => new Future.value(window.localStorage[key]);
 
   @override
   save(String json) {